requirements.txt
httpx>=0.27
tenacity>=8.3
pandas>=2.2
yfinance>=0.2
gspread>=6.0
google-auth>=2.28
python-dotenv>=1.0
pytest>=8.0
pytest-mock>=3.14
